requirements.txt
streamlit
faiss-cpu
sentence-transformers
pymupdf
python-docx
python-pptx
pandas
openai
python-dotenv
